Fix R-AStar to re-run AStar after detect wall
In the inner for-loop, when a wall is detected in the path of pathSoFar, the
code had a 'continue' statement.  This is incorrect.  It should use a
'break' statement to break out of the inner for-loop and proceed to
execute in the outer while-loop so that it can re-run AStar.  So this
commit fixes this issue.